requirements.txt
asgiref==3.8.0
astroid==3.1.0
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
dj-database-url==2.1.0
Django==5.0.4
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
filelock==3.13.3
fonttools==4.50.0
fsspec==2024.3.1
gunicorn==22.0.0
idna==3.7
isort==5.13.2
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.1.0
psutil==5.9.8
psycopg==3.1.14
psycopg-binary==3.1.14
py-cpuinfo==9.0.0
pylint==3.1.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
sympy==1.12
thop==0.1.1.post2209072238
tomlkit==0.12.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.1.37
urllib3==2.2.1
whitenoise==6.6.0
